Cleaned up code for MSVC 2008. This included modifying project settings, fixing up the nitpicky stdlib secure warnings (I didn't fix all of these, took the easy way out with #define _CRT_SECURE_NO_WARNINGS on some). Some obsolete projects removed from the solution.

+++ src/alertlib/src/AlertLib.cpp
@@ -329,7 +329,7 @@ int alert::ARGratingSpec::draw(bool useTransOnLower)
 	else
 	{	
 		// Set up standard 50:50 square wave
-		vsgObjTableSquareWave(vsgSWTABLE, vsgObjGetTableSize(vsgSWTABLE)*0.25, vsgObjGetTableSize(vsgSWTABLE)*0.75);
+		vsgObjTableSquareWave(vsgSWTABLE, (DWORD)(vsgObjGetTableSize(vsgSWTABLE)*0.25), (DWORD)(vsgObjGetTableSize(vsgSWTABLE)*0.75));
 	}
 
 	// set temporal freq
@@ -433,7 +433,7 @@ int alert::ARApertureGratingSpec::draw()
 	else
 	{	
 		// Set up standard 50:50 square wave
-		vsgObjTableSquareWave(vsgSWTABLE, vsgObjGetTableSize(vsgSWTABLE)*0.25, vsgObjGetTableSize(vsgSWTABLE)*0.75);
+		vsgObjTableSquareWave(vsgSWTABLE, (DWORD)(vsgObjGetTableSize(vsgSWTABLE)*0.25), (DWORD)(vsgObjGetTableSize(vsgSWTABLE)*0.75));
 	}
 
 	// set temporal freq
